Modify MLGameState
- Add logic for peeking
- Add MLGameState to its own file
